+++ lib/slack-ruby-client.rb
@@ -4,12 +4,14 @@ require 'slack/config'
 # Web API
 require 'faraday'
 require 'faraday_middleware'
+require 'json'
 require 'slack/web/config'
 require 'slack/web/api/error'
 require 'slack/web/faraday/response/raise_error'
 require 'slack/web/faraday/connection'
 require 'slack/web/faraday/request'
 require 'slack/web/api/endpoints'
+require 'slack/web/api/patches'
 require 'slack/web/client'
 
 # RealTime API