+++ lib/slack/web/api/patches.rb
@@ -1,0 +1,1 @@
+require 'slack/web/api/patches/chat'